+++ src/Pho/functions.php
@@ -13,6 +13,7 @@ namespace Eloquent\Phony\Pho;
 
 use Eloquent\Phony\Call\Argument\ArgumentsInterface;
 use Eloquent\Phony\Event\EventCollectionInterface;
+use Eloquent\Phony\Event\EventInterface;
 use Eloquent\Phony\Integration\Pho\PhoFacadeDriver;
 use Eloquent\Phony\Matcher\MatcherInterface;
 use Eloquent\Phony\Mock\Builder\MockBuilderInterface;
@@ -193,7 +194,7 @@ function stub($callback = null)
  *
  * @api
  *
- * @param EventCollectionInterface ...$events The events.
+ * @param EventInterface|EventCollectionInterface ...$events The events.
  *
  * @return EventCollectionInterface|null The result.
  */
@@ -209,7 +210,7 @@ function checkInOrder()
  *
  * @api
  *
- * @param EventCollectionInterface ...$events The events.
+ * @param EventInterface|EventCollectionInterface ...$events The events.
  *
  * @return EventCollectionInterface The result.
  * @throws Exception                If the assertion fails, and the assertion recorder throws exceptions.
@@ -225,7 +226,7 @@ function inOrder()
  *
  * @api
  *
- * @param mixed<EventCollectionInterface> $events The event sequence.
+ * @param mixed<EventInterface|EventCollectionInterface> $events The event sequence.
  *
  * @return EventCollectionInterface|null The result.
  */
@@ -241,7 +242,7 @@ function checkInOrderSequence($events)
  *
  * @api
  *
- * @param mixed<EventCollectionInterface> $events The event sequence.
+ * @param mixed<EventInterface|EventCollectionInterface> $events The event sequence.
  *
  * @return EventCollectionInterface The result.
  * @throws Exception                If the assertion fails, and the assertion recorder throws exceptions.
@@ -257,7 +258,7 @@ function inOrderSequence($events)
  *
  * @api
  *
- * @param EventCollectionInterface ...$events The events.
+ * @param EventInterface|EventCollectionInterface ...$events The events.
  *
  * @return EventCollectionInterface|null The result.
  * @throws InvalidArgumentException      If invalid input is supplied.
@@ -273,7 +274,7 @@ function checkAnyOrder()
  *
  * @api
  *
- * @param EventCollectionInterface ...$events The events.
+ * @param EventInterface|EventCollectionInterface ...$events The events.
  *
  * @return EventCollectionInterface The result.
  * @throws InvalidArgumentException If invalid input is supplied.
@@ -290,7 +291,7 @@ function anyOrder()
  *
  * @api
  *
- * @param mixed<EventCollectionInterface> $events The event sequence.
+ * @param mixed<EventInterface|EventCollectionInterface> $events The event sequence.
  *
  * @return EventCollectionInterface|null The result.
  * @throws InvalidArgumentException      If invalid input is supplied.
@@ -307,7 +308,7 @@ function checkAnyOrderSequence($events)
  *
  * @api
  *
- * @param mixed<EventCollectionInterface> $events The event sequence.
+ * @param mixed<EventInterface|EventCollectionInterface> $events The event sequence.
  *
  * @return EventCollectionInterface The result.
  * @throws InvalidArgumentException If invalid input is supplied.