+++ src/functions.php
@@ -13,6 +13,7 @@ namespace Eloquent\Phony;
 
 use Eloquent\Phony\Call\Argument\ArgumentsInterface;
 use Eloquent\Phony\Event\EventCollectionInterface;
+use Eloquent\Phony\Event\EventInterface;
 use Eloquent\Phony\Facade\FacadeDriver;
 use Eloquent\Phony\Matcher\MatcherInterface;
 use Eloquent\Phony\Mock\Builder\MockBuilderInterface;
@@ -191,7 +192,7 @@ function stub($callback = null)
  *
  * @api
  *
- * @param EventCollectionInterface ...$events The events.
+ * @param EventInterface|EventCollectionInterface ...$events The events.
  *
  * @return EventCollectionInterface|null The result.
  */
@@ -207,7 +208,7 @@ function checkInOrder()
  *
  * @api
  *
- * @param EventCollectionInterface ...$events The events.
+ * @param EventInterface|EventCollectionInterface ...$events The events.
  *
  * @return EventCollectionInterface The result.
  * @throws Exception                If the assertion fails, and the assertion recorder throws exceptions.
@@ -223,7 +224,7 @@ function inOrder()
  *
  * @api
  *
- * @param mixed<EventCollectionInterface> $events The event sequence.
+ * @param mixed<EventInterface|EventCollectionInterface> $events The event sequence.
  *
  * @return EventCollectionInterface|null The result.
  */
@@ -239,7 +240,7 @@ function checkInOrderSequence($events)
  *
  * @api
  *
- * @param mixed<EventCollectionInterface> $events The event sequence.
+ * @param mixed<EventInterface|EventCollectionInterface> $events The event sequence.
  *
  * @return EventCollectionInterface The result.
  * @throws Exception                If the assertion fails, and the assertion recorder throws exceptions.
@@ -255,7 +256,7 @@ function inOrderSequence($events)
  *
  * @api
  *
- * @param EventCollectionInterface ...$events The events.
+ * @param EventInterface|EventCollectionInterface ...$events The events.
  *
  * @return EventCollectionInterface|null The result.
  * @throws InvalidArgumentException      If invalid input is supplied.
@@ -271,7 +272,7 @@ function checkAnyOrder()
  *
  * @api
  *
- * @param EventCollectionInterface ...$events The events.
+ * @param EventInterface|EventCollectionInterface ...$events The events.
  *
  * @return EventCollectionInterface The result.
  * @throws InvalidArgumentException If invalid input is supplied.
@@ -288,7 +289,7 @@ function anyOrder()
  *
  * @api
  *
- * @param mixed<EventCollectionInterface> $events The event sequence.
+ * @param mixed<EventInterface|EventCollectionInterface> $events The event sequence.
  *
  * @return EventCollectionInterface|null The result.
  * @throws InvalidArgumentException      If invalid input is supplied.
@@ -305,7 +306,7 @@ function checkAnyOrderSequence($events)
  *
  * @api
  *
- * @param mixed<EventCollectionInterface> $events The event sequence.
+ * @param mixed<EventInterface|EventCollectionInterface> $events The event sequence.
  *
  * @return EventCollectionInterface The result.
  * @throws InvalidArgumentException If invalid input is supplied.