+++ test/suite/Call/Event/ReceivedExceptionEventTest.php
@@ -33,9 +33,6 @@ class ReceivedExceptionEventTest extends PHPUnit_Framework_TestCase
         $this->assertSame($this->time, $this->subject->time());
         $this->assertSame($this->exception, $this->subject->exception());
         $this->assertNull($this->subject->call());
-        $this->assertTrue($this->subject->hasEvents());
-        $this->assertSame(array($this->subject), $this->subject->allEvents());
-        $this->assertSame(1, count($this->subject));
     }
 
     public function testConstructorDefaults()
@@ -45,29 +42,11 @@ class ReceivedExceptionEventTest extends PHPUnit_Framework_TestCase
         $this->assertEquals($this->exception, $this->subject->exception());
     }
 
-    public function testIteration()
-    {
-        $this->assertSame(array($this->subject), iterator_to_array($this->subject));
-    }
-
     public function testSetCall()
     {
         $call = $this->callFactory->create();
         $this->subject->setCall($call);
 
         $this->assertSame($call, $this->subject->call());
-    }
-
-    public function testEventAt()
-    {
-        $this->assertSame($this->subject, $this->subject->eventAt());
-        $this->assertSame($this->subject, $this->subject->eventAt(0));
-        $this->assertSame($this->subject, $this->subject->eventAt(-1));
-    }
-
-    public function testEventAtFailure()
-    {
-        $this->setExpectedException('Eloquent\Phony\Event\Exception\UndefinedEventException');
-        $this->subject->eventAt(1);
     }
 }